Statehandling example for StyleEditor
