+++ src/react/components/StyleEditor.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { LocaleProvider } from '../util';
 import { Message } from './Message';
@@ -46,28 +46,13 @@ export const StyleEditor = (props) => {
     let [form] = Form.useForm();
 
     // initialize state with propvided style settings to show preview correctly and set default format as point
-    //const [state, setState] = useState({ ...props.oskariStyle });
     const fieldValuesForForm = FormToOskariMapper.createFlatFormObjectFromStyle(props.oskariStyle);
     form.setFieldsValue(fieldValuesForForm);
     const [selectedTab, setSelectedTab] = useState(props.format || 'point');
-/*
-      const formSetCallback = (valueToSet) => form.setFieldsValue(valueToSet); // callback for populating form with provided values
-      const stateSetCallback = (newState) => setState({ ...newState}); // callback for setting state of form - with this we force re-render even though state is handled in handler
-
-      props.formHandler.setCallbacks(stateSetCallback, formSetCallback);
-      props.formHandler.populateWithStyle(props.oskariStyle);
-      props.formHandler.populateWithStyle(props.formHandler.getCurrentStyle());
-*/
     const updateStyle = FormToOskariMapper.createStyleAdjuster(props.oskariStyle);
     const onUpdate = (values) => {
         // {image.shape: 3}
         const newStyle = updateStyle(values);
-        /*
-        console.log('Form triggered update:', values);
-        console.log('Original style:', props.oskariStyle);
-        console.log('Modified style:', newStyle);
-        // TODO: trigger onChange(newStyle) instead
-        */
         props.onChange(newStyle)
     };
 
@@ -92,4 +77,10 @@ export const StyleEditor = (props) => {
             </Space>
         </LocaleProvider>
     );
+};
+
+StyleEditor.propTypes = {
+    oskariStyle: PropTypes.object,
+    format: PropTypes.oneOf(['point', 'line', 'area']),
+    onChange: PropTypes.func.isRequired
 };
